+++ examples/tutorials/04_pie.py
@@ -5,6 +5,7 @@ def main():
     set_color("red")
     set_fill_color("lightyellow")
     draw_pie(200, 150, 135, 45, 100, 70)
+    draw_arc(200, 250, 135, 45, 100, 70)
     pause()
     close_graph()